Created universal test runners
